# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation about projekt SasiadWPotrzebie",
        "description": "sasiadwpotrzebie api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "local server"
        },
        {
            "url": "http://localhost:9000",
            "description": "staging server"
        },
        {
            "url": "http://localhost:9000",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/api/communes": {
            "get": {
                "tags": [
                    "Communes"
                ],
                "summary": "Retrieve all communes",
                "description": "Get all communes",
                "operationId": "a94ea9d4353d0743eacc3a91e67095ad",
                "responses": {
                    "200": {
                        "description": "List of all communes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Commune Alpha"
                                            },
                                            "district_id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-22T10:20:30Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-23T15:30:45Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/communes/{id}": {
            "get": {
                "tags": [
                    "Communes"
                ],
                "summary": "Retrieve a commune by its ID",
                "description": "Get commune by id",
                "operationId": "c4a597b9a1678543121955e440f75230",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Commune ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commune details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Commune Beta"
                                        },
                                        "district_id": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-09-22T10:20:30Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-09-23T15:30:45Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commune not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/districts/{id}/communes": {
            "get": {
                "tags": [
                    "Communes"
                ],
                "summary": "Retrieve communes by district ID",
                "description": "Get communes by district",
                "operationId": "3d22538bdce5cfaeabb9edb121b76a7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "District ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of communes in the district",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Commune Gamma"
                                            },
                                            "district_id": {
                                                "type": "integer",
                                                "example": 30
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-25T11:15:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-26T17:45:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "District not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}